[refactor] add lib/handler.js
