--- frontend/src/lib/handleArray.js
@@ -1,4 +1,0 @@
-export const isEmptyArr = (arr = []) => {
-  if (arr.length < 1) return true;
-  return false;
-};
