--- frontend/src/lib/handleString.js
@@ -1,6 +1,0 @@
-const EMPTY = '';
-
-export const isEmptyStr = (string = EMPTY) => {
-  if (string === EMPTY) return true;
-  return false;
-};
