+++ frontend/src/lib/handler.js
@@ -1,0 +1,8 @@
+export const isEmptyArr = (arr = []) => {
+  return arr.length < 1;
+};
+
+const EMPTY = '';
+export const isEmptyStr = (string = EMPTY) => {
+  return string === EMPTY;
+};
